Update financial_planning_tools.ipynb
Used the Alpaca get_bars function to get current closing prices the portfolio. Then, I reorganized the DataFrame and separated ticker data. I concatenated the ticker DataFrames and then reviewed the 1st 5 rows of the data, but for this data there is only 1 row because there is only 1 day of data